api/pyproject.toml
[project]
name = "lastfm-last-played"
version = "0.1.0"
description = "Last.fm Last Played Songs API"
requires-python = ">=3.12"
dependencies = [
    "flask==2.3.3",
    "gunicorn==23.0.0",
    "flask-cors",
    "requests",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-flask",
    "opentelemetry-instrumentation",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp"
]

[project.optional-dependencies]
dev = [
    "pylint",
    "autopep8",
    "pytest",
    "faker",
    "pytest-cov"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lastfm_last_played"]
